Desktop/code bot/requirements.txt
requests>=2.31.0,<3
fastapi>=0.110.0,<1
uvicorn[standard]>=0.27.0,<1
python-dotenv>=1.0.1,<2
starlette>=0.36.0,<1
